refactor: improve general api
